+++ packages/data-model/src/converter/AcDbDxfParser.ts
@@ -1,7 +1,9 @@
 import DxfParser, { ParsedDxf } from '@mlightcad/dxf-json'
 
+// Please don't modify the following two lines to import from '../database' and
+// import from '../misc' so that treeshaking can include classes really needed.
 import { AcDbDwgVersion } from '../database/AcDbDwgVersion'
-import { AcDbCodePage,dwgCodePageToEncoding } from '../misc/AcDbCodePage'
+import { AcDbCodePage, dwgCodePageToEncoding } from '../misc/AcDbCodePage'
 
 /**
  * Extracts DXF version and code page from an ArrayBuffer containing the DXF data.