+++ packages/libdxfrw-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libdxfrw-converter
 
+## 3.1.1
+
+### Patch Changes
+
+- feat: add api to support redrawing the database
+- Updated dependencies
+  - @mlightcad/data-model@1.3.1
+
 ## 3.1.0
 
 ### Minor Changes